// react-java/frontend/src/types/api/types.gen.ts
export type LoginRequestDto = {
    username?: string;
    password?: string;
};

export type LoginResponseDto = {
    accessToken?: string;
    accessExpireIn?: number;
    refreshToken?: string;
    refreshExpireIn?: number;
};
// ...
export type RoleAuthInfoDto = {
    coId?: string;
    creUsrId?: string;
    updUsrId?: string;
    creDt?: string;
    updDt?: string;
    useFlg?: string;
    procFlag?: string;
    roleId?: string;
    pgmId?: string;
    permId?: string;
    pgmCd?: string;
    pgmTpCd?: string;
    permCd?: string;
    activeYn?: string;
};